rtlamr-mqttexporter/requirements.txt
paho-mqtt==1.5.0
docker==4.1.0
